Task 2.1.5: unify state_factory scalar dtypes with env.reset_env
Wrap selected_player and timestep in jnp.int32(...) in state_factory.
This matches what factoriax/levels.py:686-687 emits in env.reset_env,
so a state_factory-built state has bit-identical pytree shape to a
reset_env state. Two-line fix; the rest of state_factory's leaves
already matched reset_env's dtypes (audited leaf-by-leaf in 2.2).

Impact on the four Task 2.2 migrated tests:

| Test                                    | Before 2.1.5 | After |
|-----------------------------------------|--------------|-------|
| test_delta_resets_each_step             |    6.76s     | 6.61s |
| test_step_consumes_and_resets           |    6.73s     | 0.01s |
| test_params_kwarg_populates_scheme      |    0.03s     | 0.03s |
| test_omitting_params_leaves_scheme_none |    0.03s     | 0.02s |
| Total (four migrated)                   |   13.55s     | 6.67s |

test_step_consumes_and_resets — the lab-state test that previously
retraced — now hits cache because the state_factory-built lab state
shares its pytree shape with the canonical fixture's reset state.
That's a 77% reduction on the four migrated tests vs the original
29.66s baseline.

Full suite: 1355 passed in 733.15s (was 775.81s after 2.2; ~42s
saved). The wider improvement is consistent with the change
unblocking cache hits anywhere state_factory states cross paths with
env.reset_env states inside a single test's jit_step_fn calls.

No assertion changed; all 1355 tests still pass with coverage at 38%.

+++ tests/conftest.py
@@ -314,8 +314,13 @@ def state_factory():
                 if player_inventory is not None
                 else jnp.zeros(inv_shape, dtype=jnp.int16)
             ),
-            selected_player=selected_player,
-            timestep=timestep,
+            # Match env.reset_env's pytree shape: factoriax/levels.py:686-687
+            # emits these as jnp.int32(0); a Python-int leaf here would
+            # force jax.jit(env.step_env) to retrace whenever a test feeds
+            # a state_factory state through the canonical_env_8x8_1p
+            # fixture's shared step path.
+            selected_player=jnp.int32(selected_player),
+            timestep=jnp.int32(timestep),
             items_mined=(
                 items_mined
                 if items_mined is not None
